Separate build logic for different workspaces
